[FIX] ces_events shortcode description show/hide
Summary: work around JS event handling; prevent accessibility
complications.

In 3f268d2b, accordion.js's event listener for the
"Description and Objectives" button expects the event target
to be an element having the class "Accordion-trigger", which
is not true of the .Accordion-trigger's child elements,
which would very reasonably get clicked to operate the button's
functionality.

Because spans inside a button can get tricky with accessibility,
we are opting to just remove the child elements from our buttons.
This means that accordion.js still does not handle the case when
.Accordion-trigger's child element(s) are clicked.

Co-authored-by: ecosio ecosio@odu.edu