[FIX] Allow <tr> to contain <textareas> in TNP editor
The height=100 for each tr conflicted with a dynamic
height value added to a focused textarea by TNP's JS.
Because the row as 100px and the textarea 400px, we
got a visual bug.

This commit also removes additional styles that weren't
strictly necessary. It seems TNP does a good job of
of styling the rows without the additions.